+++ src/main/java/org/dive2025/qdeep/domain/file/dto/request/LoadImagesRequest.java
@@ -1,0 +1,4 @@
+package org.dive2025.qdeep.domain.file.dto.request;
+
+public record LoadImagesRequest(Long boardId) {
+}
